Refactor truss and integration_tests
truss has been refactored into main.go and generator/generator.go.
Where main.go deals with working directory, GOPATH, and argument
parsing and sanitation and generator.go actually exec's protoc and go
build to generate and build the microservice.

integration_tests were refactored to use `testing.T` to do logging
rather than logrus. As well as be more readable. It can also be called
with `$ go test -tags=integration -clean` in order to remove the
generated test files and nothing else.

Eventual goals are to unit test generator.go's functions and
eventually move the integration_tests to the generator package and test
by calling functions rather than execing the truss binary
